+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## [Unreleased]
 
-## [0.1.1] - 2025-01-21
+## [0.1.2] - 2025-01-28
+
+- Update usage on sub-commands.
+- Allow an optional description to be included when calling .cmdadd().
+
+## [0.1.1] - 2025-01-28
 
 - Add default "help" sub-command to all commands.
 - Allow for custom command lookups by overriding SubOptParser#[].